requirements.txt
# Core Framework - FastAPI (keeping for performance, but adapting to requirements)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database - MongoDB as per HLD/LDL/PRM requirements
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0  # Sync MongoDB driver
beanie==1.24.0  # ODM for MongoDB with Pydantic
dnspython==2.4.2  # For MongoDB Atlas connections

# Redis for caching and real-time features
redis==5.0.1
aioredis==2.0.1

# LangChain & AI (as per requirements) - Fixed version compatibility
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.15
langgraph==0.2.34
langserve==0.3.1

# EURI AI Integration (official SDK)
euriai  # Official EURI AI Python SDK

# Additional AI Model integrations (as per PRM requirements)
openai==1.6.1  # For OpenAI integration mentioned in requirements
stability-sdk==0.8.1  # For Stability AI (visual generation)

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.25.2

# File Processing
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.1.0

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# CORS (handled by FastAPI built-in middleware)

# WebSocket support
websockets==12.0

# Background tasks
celery==5.3.4

# Email
fastapi-mail==1.4.1

# Monitoring
prometheus-client==0.19.0

# Additional utilities
python-slugify==8.0.1
jinja2==3.1.2
